pyproject.toml
[tool.poetry]
name = "together-api"
version = "0.1.0"
description = ""
authors = ["Mykyta Kasianenko <mka@empat.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
psycopg2 = "^2.9.9"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
python-socketio = {extras = ["asyncio"], version = "^5.11.4"}
redis = "^5.0.8"
pytest-asyncio = "^0.24.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
